3LA 폴더 생성 및 리펙토링
