+++ homework/songsamuel/backend/db/todoitems.js
@@ -1,0 +1,18 @@
+export default [
+  {
+    id: 1,
+    userId: 1,
+    title: "할일1",
+    doneAt: null,
+    createdAt: "2021-08-01",
+    updatedAt: "2021-08-01",
+  },
+  {
+    id: 2,
+    userId: 2,
+    title: "할일2",
+    doneAt: null,
+    createdAt: "2021-08-01",
+    updatedAt: "2021-08-01",
+  },
+];
