+++ homework/songsamuel/backend/db/users.js
@@ -1,0 +1,16 @@
+export default [
+  {
+    id: 1,
+    email: "thdtkandpf@naver.com",
+    password: "aaaa4321",
+    role: "학생",
+    name: "송사무엘",
+  },
+  {
+    id: 2,
+    email: "ssong@naver.com",
+    password: "aaaa4321",
+    role: "학생",
+    name: "쏭쏭쏭",
+  },
+];
