+++ homework/songsamuel/backend/middlewares/leaveLogMiddleware.js
@@ -1,0 +1,7 @@
+export default (req, res, next) => {
+  console.log(`
+        ${req.method} ${req.url} [${new Date().toISOString()}] ${
+    req.headers.referer
+  }`);
+  next();
+};
